% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc-pnc.R
\name{calc.pnc}
\alias{calc.pnc}
\title{calc.pnc}
\usage{
calc.pnc(city = "nyc", method = "complete", nrpts = 100, rescale = 2)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{complete, ward, or skater, to be compared to k-means}

\item{rescale}{order of polynomial rescaling of peak heights}

\item{nprts}{number of random trials used to calculate probabilities}
}
\value{
data.frame and also dumps results to text file
}
\description{
Calculates probabilities of peak heights as returned from the num.clusts ()
function.  If is.na (dat), then the table is called internally. To generate
useful statistics, this function should be run with a VERY large number of
repeats (for example, 100,000), which will take a LONG time!
}

