% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust-sig.R
\name{clust.sig}
\alias{clust.sig}
\title{clust.sig}
\usage{
clust.sig(city = "nyc", method = "complete", xmax = 36)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{= (ward, k-means, complete)}

\item{xmax}{maximum number of clusters to be analysed and plotted}
}
\value{
statistical summary as screen dump
}
\description{
Uses the output of calc.pnc, which calculates the probability of observing a
given peak height for a given number of clusters. This routine in turn uses
the output of num.clusts.
}

