% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare-dists.R
\name{compare.dists}
\alias{compare.dists}
\title{compare.dists}
\usage{
compare.dists(city = "nyc", method = "complete", xmax = 36,
  ycut = c(1.2, 1.3))
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{= (ward, k-means, complete)}

\item{xmax}{maximum number of clusters to be analysed and plotted}
}
\value{
data.frame
}
\description{
Calculates mean ratio of observed to expected intra-cluster distances
calculated out to given number of clusters.
}

