% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-clusters.R
\name{get.clusters}
\alias{get.clusters}
\title{get-clusters}
\usage{
get.clusters(city = "nyc", method = "ward", trial = "")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{= (ward, k-means, complete)}

\item{trial}{is appended to filename (used to generate multiple k-means
results).}
}
\value{
nothing (dumps files to dir)
}
\description{
Produces .csv files with cluster memberships across a range of numbers of
clusters (1-max_clust_size=100), along with corresponding average cluster
diameters (in km).
}

