% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-members.R
\name{get.members}
\alias{get.members}
\title{get-members}
\usage{
get.members(city = "nyc", nc = 8, method = "ward", details = FALSE)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{nc}{number of clusters}

\item{method}{= (ward, k-means, complete)}

\item{details}{prints statistics of cluster sizes}
}
\value{
a vector of 2 columns for (from, to) data, each containing cluster
numbers for every bike station.
}
\description{
Creates clusters and then applies spatial constraint to reallocate stray
points to neighbouring clusters. "method" can be ward, complete, or k-means.
All methods other than k-means are passed to hclust.
}

