% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-partition-neighbours.R
\name{get.partition.neighbours}
\alias{get.partition.neighbours}
\title{get.partition.neighbours}
\usage{
get.partition.neighbours(city = "nyc", method = "complete", nc = c(11,
  15), dir = "from", plot = FALSE)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{complete, ward, or skater, to be compared to k-means}

\item{nc}{list of 2 elements with nc[1]>nc[2], so that routine can identify
clusters in nc[1] that split to form clusters in nc[2]}

\item{dir}{to or from}

\item{plot}{enables visual inspection of the spatial arrangements.}
}
\value{
$nbs=list of neighbouring clusters in nc[1] that are subsequently
partitioned in nc[2], $sizes=corresponding sizes
}
\description{
Identifies the clusters in min (nc) that split to form separate clusters in
max (nc). For each cluster, a list element is returned identifying the groups
into which each connected group is subsequently partitioned.
}

