% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-skater-groups.R
\name{get.skater.groups}
\alias{get.skater.groups}
\title{get.skater.groups}
\usage{
get.skater.groups(city = "nyc", dirf = "from", max.groups = 50)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{max.groups}{only seems to work with skater up to around 50}

\item{dir}{= ("from", "to")}
}
\value{
nothing (dumps files to dir)
}
\description{
Produces .csv files with cluster memberships across a range of numbers of
clusters (1-max.groups=50).
}

