% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-total-dist.R
\name{get.total.dist}
\alias{get.total.dist}
\title{get.total.dist}
\usage{
get.total.dist(city = "nyc")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}
}
\value{
total distance in km
}
\description{
Calculates total distance ridden for given city
}

