% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num-clusts.R
\name{num.clusts}
\alias{num.clusts}
\title{num.clusts}
\usage{
num.clusts(city = "nyc", plot = FALSE, method = "complete")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{complete, ward, or skater, to be compared to k-means}
}
\value{
data.frame
}
\description{
Calculates peak height and G-values as a function of numbers of clusters, to
determine the number of clusters that should be used for each of the four
data series. For plot=TRUE, it also reads the table of the same values (from
"results_prob_m.txt") to plot the probabilities of the observed peak heights,
as calcualted from calc.pnc (). (The latter routine itself requires, however,
the initial output of num.clusts.)
}
\details{
These p-values are then used to determine the appropriate number of (most
highly significant) peaks to be used for analysing each series in the main
clust.sig() routine.
}

