% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poly-rescale.R
\name{poly.rescale}
\alias{poly.rescale}
\title{poly.rescale}
\usage{
poly.rescale(xm, rescale = 2)
}
\arguments{
\item{xm}{input series (as matrix for multiple rescaling)}

\item{rescale}{polynomial order of rescaling (can be > default of 2)}
}
\value{
vector of rescaled series
}
\description{
Applies quadratic regression to input series to rescale peak values between
-1 and 1
}

