% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clust-sig.R
\name{clust.sig}
\alias{clust.sig}
\title{clust.sig}
\usage{
clust.sig(city = "nyc", method = "complete", rescale = 2)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{method}{= (ward, k-means, complete)}

\item{rescale}{= order of polynomial re-scaling of T-values (2 or 3)}
}
\value{
statistical summary as screen dump
}
\description{
Statistically summarises clusters as screen dump, and also dumps brief text
files to be used in the C++ random hierarchies routine. These just contain
the positions of local peaks (in terms of numbers of clusters), and numbers
of contiguous and non-contiguous groups in which new partitions are formed at
each stage.  Uses the output of calc.pnc, which calculates the probability of
observing a given peak height for a given number of clusters.
}

